requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
tiktoken>=0.5.0
